skills/ocean_debugging/requirements.txt
# Ocean Debugging Agent Dependencies

# Web Framework (Phase 2)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Core
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Async
aiohttp>=3.9.0

# Data Clients
psycopg2-binary>=2.9.9          # Redshift (PostgreSQL)
clickhouse-driver>=0.2.7        # ClickHouse (SigNoz)
simple-salesforce>=1.12.0       # Salesforce API
requests>=2.31.0                # HTTP clients

# LLM
anthropic>=0.42.0               # Claude API (Sonnet 4.5 support)
openai>=1.59.0                  # Azure OpenAI fallback

# CLI
rich>=13.0.0                    # Terminal formatting
click>=8.1.0                    # CLI framework

# Logging
python-json-logger>=2.0.7       # JSON structured logging

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
